+++ core/src/main/java/com/ibasco/agql/core/transport/pool/NettyChannelPoolFactory.java
@@ -17,7 +17,7 @@
 package com.ibasco.agql.core.transport.pool;
 
 import com.ibasco.agql.core.transport.NettyChannelFactory;
-import com.ibasco.agql.core.util.GlobalOptions;
+import com.ibasco.agql.core.util.GeneralOptions;
 import com.ibasco.agql.core.util.Options;
 import io.netty.channel.pool.ChannelPoolHandler;
 
@@ -46,7 +46,7 @@ abstract public class NettyChannelPoolFactory {
         final Options options = Objects.requireNonNull(channelFactory, "Channel factory must not be null").getOptions();
         this.channelFactory = channelFactory;
         this.channelPoolHandler = new DefaultChannelPoolHandler(channelFactory.getBootstrap());
-        this.channelHealthChecker = options.getOrDefault(GlobalOptions.POOL_CHANNEL_HEALTH_CHECKER);
+        this.channelHealthChecker = options.getOrDefault(GeneralOptions.POOL_CHANNEL_HEALTH_CHECKER);
     }
 
     /**