+++ core/src/main/java/com/ibasco/agql/core/util/Concurrency.java
@@ -110,7 +110,7 @@ public final class Concurrency {
     public static boolean shutdown(ExecutorService executorService) {
         if (executorService == null)
             return false;
-        return shutdown(executorService, GlobalOptions.CLOSE_TIMEOUT.getDefaultValue(), TimeUnit.MILLISECONDS);
+        return shutdown(executorService, GeneralOptions.CLOSE_TIMEOUT.getDefaultValue(), TimeUnit.MILLISECONDS);
     }
 
     /**